Replace old executer by new one.
