+++ pytrip/tripexecuter/field.py
@@ -17,162 +17,110 @@
 #    along with PyTRiP98.  If not, see <http://www.gnu.org/licenses/>.
 #
 """
-TODO: documentation here.
+Field() objects here, which can be passed to a Plan() object.
 """
-from pytrip.error import InputError
-from pytrip.tripexecuter.pytripobj import pytripObj
+
+import uuid
+import logging
+
+logger = logging.getLogger(__name__)
 
 
-class Field(pytripObj):
-    def __init__(self, name):
-        self.save_fields = ["name", "gantry", "couch", "fwhm", "rasterstep", "doseextension", "contourextension",
-                            "zsteps", "projectile", "target"]
-        self.name = name
-        self.gantry = 0.0
-        self.couch = 0.0
-        self.fwhm = 5.0
-        self.rasterstep = [2, 2]
-        self.doseextension = 1.2
-        self.contourextension = 0.6
-        self.rasterfile = None
+class Field():
+    """ One or more Field() object, which then can be added to a Plan() object.
+    :params str basename: basename of field without file extension (input or output will be suffixed with
+    proper file extension)
+    """
 
-        self.zsteps = 1.0
-        self.projectile = 'C'
-        self.target = []
+    # list of projectile name - charge and most common isotope.
+    _projectile_defaults = {"H": (1, 1),
+                            "He": (2, 4),
+                            "Li": (3, 7),
+                            "C": (6, 12),
+                            "O": (8, 16),
+                            "Ne": (10, 20),
+                            "Ar": (16, 40)}
+
+    def __init__(self, basename=""):
+        self.__uuid__ = uuid.uuid4()  # for uniquely identifying this field
+        # basename of the field (i.e. without file extension).
+        # Any input/output process will be suffixed with .rst
+        self.basename = basename
+        self.number = 1  # Field number. First field must be 1
+        self.gantry = 0.0  # TRiP98 angles assumed here.
+        self.couch = 0.0  # TRiP98 angles assumed here.
+        self.fwhm = 4.0  # in [mm]
+        self.raster_step = [2, 2]  # spot width in [mm]
+        self.dose_extension = 1.2  # see TRiP98 field / doseext() documentation
+        self.contour_extension = 0.6  # see TRiP98 field / contourext() documentation
+
+        # the field / command in TRiP98 may take an external rst file as an input, and will then
+        # instead of optimization do a foward calculation based on this file.
+        # The filename will be constructed from the basename given for this field.
+        # However, we need to tell, that we provide the rst.
+        self.use_raster_file = False
+
+        self.zsteps = 1.0  # in [mm]
+        self.projectile = 'C'  # see also self._projectile_defaults
+        self.projectile_a = '12'  # Number of nucleons in projectile. If None, default will be used.
+
+        # isocenter holds the [x,y,z] coordinates of the isocenter/target in [mm].
+        # This is used for the field / target() option of the TRiP98
+        # It can be used to override the automatically calculated isocenter from TRiP98.
+        self.isocenter = []
+
         self.selected = False
 
-    def get_name(self):
-        return self.name
+    def __str__(self):
+        return self._print()
 
-    def set_rasterfile(self, path):
-        self.rasterfile = path
+    def _print(self):
+        """ Pretty print all attributes.
+        """
+        out = "\n"
+        out += "   Field {:d} '{:s}'\n".format(self.number, self.basename)
+        out += "----------------------------------------------------------------------------\n"
+        out += "|  UUID                         : {:s}\n".format(str(self.__uuid__))
+        out += "|  Projectile                   : {:s}-{:s}\n".format(self.projectile,
+                                                                      str(self.projectile_a))
+        out += "|  Couch angle                  : {:.2f} deg\n".format(self.couch)
+        out += "|  Gantry angle                 : {:.2f} deg\n".format(self.gantry)
+        out += "|\n"
+        out += "|  Spot size (FWHM              : {:.2f} mm\n".format(self.fwhm)
+        out += "|  Raster step size (x,y)       : {:.2f}, {:.2f} mm\n".format(self.raster_step[0],
+                                                                              self.raster_step[1])
+        out += "|  Z-steps                      : {:.2f} mm\n".format(self.zsteps)
+        out += "|\n"
+        out += "|  Dose extension               : {:.2f}\n".format(self.dose_extension)
+        out += "|  Contour extension            : {:.2f}\n".format(self.contour_extension)
+        out += "|  Use external .rst file       : {:s}\n".format(str(self.use_raster_file))
+        if self.isocenter:
+            out += "|  Isocenter (x,y,z)            : {:.2f} {:.2f} {:.2f} mm\n".format(self.isocenter[0],
+                                                                                        self.isocenter[1],
+                                                                                        self.isocenter[2])
+        else:
+            out += "|  Isocenter (x,y,z)            : (not set)\n"
+        out += "----------------------------------------------------------------------------\n"
+        return out
 
-    def get_rasterfile(self):
-        return self.rasterfile
+    def set_isocenter_from_string(self, isocenter_str):
+        """ Override the automatically determined isocenter from TRiP98.
+        :params str isocenter_str: x,y,z coordinates of the isocenter/target in [mm] in a comma delimted string.
+        such as "123.33,158.4,143.5".
+        If and empty string is provided, then the isocenter is unset, and TRiP98 will calculate it itself.
 
-    def set_name(self, name):
-        self.name = name
+        Following the target() option in the TRiP98 field command, one can specify
+        the isocenter of the target. If not used, TRiP98 will determine the isocenter
+        from the target Voi provided.
 
-    def is_selected(self):
-        return self.selected
-
-    def toggle_selected(self):
-        self.selected = not self.selected
-
-    def get_gantry(self):
-        return self.gantry
-
-    def set_gantry(self, angle):
-        if type(angle) is str and not len(angle):
-            angle = 0
-        try:
-            angle = (float(angle) + 360) % 360
-            if angle < 0 or angle > 360:
-                raise Exception()
-            self.gantry = angle
-
-        except Exception:
-            raise InputError("Gantry angle shoud be a " "number between 0 and 360")
-
-    def get_couch(self):
-        return self.couch
-
-    def get_rasterstep(self):
-        return self.rasterstep
-
-    def set_rasterstep(self, a, b):
-        try:
-            a = float(a)
-            b = float(b)
-            if a < 0 or b < 0:
-                raise Exception()
-            self.rasterstep = [a, b]
-        except Exception:
-            raise InputError("Rastersteps should be " "larger than 0 and numbers")
-
-    def set_couch(self, angle):
-        if type(angle) is str and not len(angle):
-            angle = 0
-        try:
-            angle = (float(angle) + 360) % 360
-            if angle < 0 or angle > 360:
-                raise Exception()
-            self.couch = angle
-
-        except Exception:
-            raise InputError("Couch angle shoud be " "a number between 0 and 360")
-
-    def set_doseextension(self, doseextension):
-        try:
-            doseextension = float(doseextension)
-            if doseextension < 0:
-                raise Exception()
-            self.doseextension = doseextension
-        except Exception:
-            raise InputError("Doseextension should be larger 0")
-
-    def get_doseextension(self):
-        return self.doseextension
-
-    def set_contourextension(self, contourextension):
-        try:
-            contourextension = float(contourextension)
-            if contourextension < 0:
-                raise Exception()
-            self.contourextension = contourextension
-        except Exception:
-            raise InputError("Contourextension should be larger 0")
-
-    def get_contourextension(self):
-        return self.contourextension
-
-    def set_zsteps(self, zsteps):
-        try:
-            zsteps = float(zsteps)
-            if zsteps < 0:
-                raise Exception()
-            self.zsteps = zsteps
-        except Exception:
-            raise InputError("ZSteps should be larger 0")
-
-    def get_zsteps(self):
-        return self.zsteps
-
-    def get_fwhm(self):
-        return self.fwhm
-
-    def set_fwhm(self, fwhm):
-        try:
-            fwhm = float(fwhm)
-            if fwhm <= 0:
-                raise Exception()
-            self.fwhm = fwhm
-
-        except Exception:
-            raise InputError("Fwhm shoud be a number and larger than 0")
-
-    def get_projectile(self):
-        return self.projectile
-
-    def set_projectile(self, projectile):
-
-        if projectile not in ['H', 'C', 'O', 'Ne']:
-            raise InputError("Projectile not allowed")
-        self.projectile = projectile
-
-    def get_target(self):
-        return self.target
-
-    def set_target(self, target):
-        if len(target) is 0:
-            self.target = []
+        """
+        if len(isocenter_str) is 0:
+            self.isocenter = []
             return
-        target = target.split(",")
-        if len(target) is 3:
+        _target = isocenter_str.split(",")
+        if len(_target) is 3:
             try:
-                self.target = [float(target[0]), float(target[1]), float(target[2])]
+                self.isocenter = [float(_target[0]), float(_target[1]), float(_target[2])]
                 return
             except Exception:
-                # TODO fix that !
-                pass
-        raise InputError("Target should be empty " "or in the format x,y,z")
+                logger.error("Expected a 'X,Y,Z' formatted string for Field().set_isocenter_from_string")
